Crate angular

Reexports

pub use Angle::{Radians, Degrees};

Enums

Angle
An angle.

Functions

acos
Compute the arccosine of a number. Return value is in the range of [0, π] rad or None if the number is outside the range [-1, 1].

asin
Compute the arcsine of a number. Return value is in the range of [-π/2, π/2] rad or None if the number is outside the range [-1, 1].

atan
Compute the arctangent of a number. Return value is in the range of [-π/2, π/2] rad.

atan2
Compute the four quadrant arctangent of y and x.

mean_angle
Compute the approximate mean of a list of angles by averaging the Cartesian coordinates of the angles on the unit circle. Return the normalized angle.