Enum angular::Angle

pub enum Angle<T = f64> {
    Radians(T),
    Degrees(T),
}

An angle.

Might be a value in degrees or in radians.

Variants

Radians

Degrees

Methods

impl<T: NumCast> Angle<T>

fn in_radians(self) -> T

Yield the value encoded in radians.

fn in_degrees(self) -> T

Yield the value encoded in degrees.

fn eighth() -> Angle<T>

An angle of 45°.

fn quarter() -> Angle<T>

An angle of 90° (right angle).

fn half() -> Angle<T>

An angle of 180° (straight).

fn full() -> Angle<T>

An angle of 360° (perigon).

impl<T: Copy + Num + NumCast + PartialOrd> Angle<T>

fn normalized(self) -> Self

Create a new angle by normalizing the value into the range of [0, 2π) rad.

Examples

let alpha = Degrees(-90.0f64).normalized();
assert!((alpha.in_degrees() - 270.0).abs() < 1.0e-10);

let beta = Radians(2.0 * PI).normalized();
assert!((beta.in_radians() - 0.0).abs() < 1.0e-10);

impl<T: Float> Angle<T>

fn min_dist(self, other: Angle<T>) -> Angle<T>

Computes the minimal unsigned distance between two normalized angles. Returns an angle in the range of [0, π] rad.

let distance = Degrees(345.0).min_dist(Degrees(15.0));
assert!((distance.in_degrees() - 30.0) < 1.0e-10);

impl<T: Signed> Angle<T>

fn abs(self) -> Self

Compute the absolute angle.

impl<T: Float + NumCast> Angle<T>

fn sin(self) -> T

Compute the sine of the angle.

fn cos(self) -> T

Compute the cosine of the angle.

fn tan(self) -> T

Compute the tangent of the angle.

fn sin_cos(self) -> (T, T)

Simultaneously compute the sine and cosine of the number, x. Return (sin(x), cos(x)).